Separate place stat vars related APIs in a separate file
