+++ internal/server/places.go
@@ -17,7 +17,6 @@ package server
 import (
 	"context"
 	"fmt"
-	"sort"
 	"strings"
 
 	"encoding/json"
@@ -190,91 +189,4 @@ func (s *Server) GetLocationsRankings(ctx context.Context,
 		}
 	}
 	return &pb.GetLocationsRankingsResponse{Payload: results}, nil
-}
-
-// GetPlaceStatsVar implements API for Mixer.GetPlaceStatsVar.
-func (s *Server) GetPlaceStatsVar(
-	ctx context.Context, in *pb.GetPlaceStatsVarRequest) (
-	*pb.GetPlaceStatsVarResponse, error) {
-
-	req := pb.GetPlaceStatVarsRequest{Dcids: in.GetDcids()}
-	resp, err := s.GetPlaceStatVars(ctx, &req)
-	if err != nil {
-		return nil, err
-	}
-	out := pb.GetPlaceStatsVarResponse{Places: map[string]*pb.StatsVars{}}
-	for dcid, statVars := range resp.Places {
-		out.Places[dcid] = &pb.StatsVars{StatsVars: statVars.StatVars}
-	}
-	return &out, nil
-}
-
-// GetPlaceStatVars implements API for Mixer.GetPlaceStatVars.
-func (s *Server) GetPlaceStatVars(
-	ctx context.Context, in *pb.GetPlaceStatVarsRequest) (
-	*pb.GetPlaceStatVarsResponse, error) {
-	dcids := in.GetDcids()
-	if len(dcids) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "Missing required arguments: dcid")
-	}
-	rowList := buildPlaceStatsVarKey(dcids)
-	dataMap, err := bigTableReadRowsParallel(ctx, s.btTable, rowList,
-		func(dcid string, jsonRaw []byte) (interface{}, error) {
-			var data PlaceStatsVar
-			err := json.Unmarshal(jsonRaw, &data)
-			if err != nil {
-				return nil, err
-			}
-			return data.StatVarIds, nil
-		}, nil)
-	if err != nil {
-		return nil, err
-	}
-	resp := pb.GetPlaceStatVarsResponse{Places: map[string]*pb.StatVars{}}
-	for _, dcid := range dcids {
-		resp.Places[dcid] = &pb.StatVars{StatVars: []string{}}
-		if dataMap[dcid] != nil {
-			resp.Places[dcid].StatVars = dataMap[dcid].([]string)
-		}
-	}
-	return &resp, nil
-}
-
-// keysToSlice stores the keys of a map in a slice.
-func keysToSlice(m map[string]bool) []string {
-	s := make([]string, len(m))
-	i := 0
-	for k := range m {
-		s[i] = k
-		i++
-	}
-	sort.Strings(s)
-	return s
-}
-
-// GetPlaceStatVarsUnion implements API for Mixer.GetPlaceStatVarsUnion.
-func (s *Server) GetPlaceStatVarsUnion(
-	ctx context.Context, in *pb.GetPlaceStatVarsUnionRequest) (
-	*pb.GetPlaceStatVarsUnionResponse, error) {
-	dcids := in.GetDcids()
-	if len(dcids) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "Missing required arguments: dcid")
-	}
-	resp, err := s.GetPlaceStatVars(ctx, &pb.GetPlaceStatVarsRequest{Dcids: dcids})
-	if err != nil {
-		return nil, err
-	}
-	places := resp.GetPlaces()
-	// Get union of the statvars for each place.
-	set := map[string]bool{}
-	for _, statVars := range places {
-		for _, dcid := range statVars.GetStatVars() {
-			set[dcid] = true
-		}
-	}
-	return &pb.GetPlaceStatVarsUnionResponse{
-		StatVars: &pb.StatVars{
-			StatVars: keysToSlice(set),
-		},
-	}, nil
 }
